Trims down hyrax collections helper
* moves  collection_permission_template_form_for into application helper
* removes copy of hyrax collections helper

It turns out collection_permission_template_form_for was from an
unreleased version of the helper, so we can just define it as a normal
view helper, without redefining the hyrax collections_helper.
